+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Not released
 
+- Added filterable prop to TimeSeriesWidget [#808](https://github.com/CartoDB/carto-react/pull/808)
 - Fix dataSources store type [#807](https://github.com/CartoDB/carto-react/pull/807)
 
 ## 2.3